+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.8
+
+- fix web-images-path for linux, macos
+- add argument for custom web-port
+
+
 ## 1.0.7
 
 - fix web-host